+++ build.washingtonpost.com/services/getDocByPathName.js
@@ -1,27 +1,13 @@
 import fs from "fs";
-import globby from "globby";
 import path from "path";
 import matter from "gray-matter";
 import { serialize } from "next-mdx-remote/serialize";
 import { remarkMdxCodeMeta } from "remark-mdx-code-meta";
 import remarkGfm from "remark-gfm";
 
-const cache = new Map();
-
 export async function getHeadings(input) {
-  // if in cache already, return it
-  let frontMatter = null;
-
-  if (cache.has(input)) {
-    frontMatter = cache.get(input);
-  } else {
-    // otherwise, read the file and cache it
-    const fileData = fs.readFileSync(`docs/${input}.mdx`);
-    const frontMatter = matter(fileData);
-
-    // add to cache
-    cache.set(input, frontMatter);
-  }
+  const fileData = fs.readFileSync(`docs/${input}.mdx`);
+  const frontMatter = matter(fileData);
 
   // Get each line individually, and filter out anything that
   // isn't a heading.
@@ -46,23 +32,13 @@ export async function getHeadings(input) {
  * gets markdown content given a path
  */
 export const getDocByPathName = async (input) => {
-  // if in cache already, return it
   let frontMatter = null;
 
-  if (cache.has(input)) {
-    frontMatter = cache.get(input);
-  } else {
-    // otherwise, read the file and cache it
+  const [directory, fileName] = input.split("/");
+  const slug = path.join(`docs/${directory}`, `${fileName}.mdx`);
+  const fileData = fs.readFileSync(slug);
 
-    const [directory, fileName] = input.split("/");
-    const slug = path.join(`docs/${directory}`, `${fileName}.mdx`);
-    const fileData = fs.readFileSync(slug);
-
-    frontMatter = matter(fileData);
-
-    // add to cache
-    cache.set(input, frontMatter);
-  }
+  frontMatter = matter(fileData);
 
   const source = await serialize(frontMatter.content, {
     scope: frontMatter.data,
@@ -77,18 +53,10 @@ export const getDocByPathName = async (input) => {
 export const getResource = async (input) => {
   let frontMatter = null;
 
-  if (cache.has(input)) {
-    frontMatter = cache.get(input);
-  } else {
-    // otherwise, read the file and cache it
-    const [directory, category, fileName] = input.split("/");
-    const slug = path.join(`docs/${directory}/${category}`, `${fileName}.mdx`);
-    const fileData = fs.readFileSync(slug);
-    frontMatter = matter(fileData);
-
-    // add to cache
-    cache.set(input, frontMatter);
-  }
+  const [directory, category, fileName] = input.split("/");
+  const slug = path.join(`docs/${directory}/${category}`, `${fileName}.mdx`);
+  const fileData = fs.readFileSync(slug);
+  frontMatter = matter(fileData);
 
   const source = await serialize(frontMatter.content, {
     scope: frontMatter.data,