+++ CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.12.0](https://github.com/washingtonpost/wpds-ui-kit/compare/v0.11.2...v0.12.0) (2022-08-03)
+
+
+### Bug Fixes
+
+* **docs:** EditInGithub alignment bug ([d18cbbd](https://github.com/washingtonpost/wpds-ui-kit/commit/d18cbbdacc96fc5b3e38e567a40177d5d5703e1d))
+* float label on controlled input value changes ([949f7df](https://github.com/washingtonpost/wpds-ui-kit/commit/949f7dfd19d2610bb904085dcc63231ea66c3de9))
+
+
+### Features
+
+* move docs site into ui-kit to create monorepo ([#141](https://github.com/washingtonpost/wpds-ui-kit/issues/141)) ([7f0b38c](https://github.com/washingtonpost/wpds-ui-kit/commit/7f0b38c73e3708fe3ab7c302405036e0fa80855c))
+* upgrade to stitches 1.2.8, fix * imports in theme ([8cbff76](https://github.com/washingtonpost/wpds-ui-kit/commit/8cbff76b14fc99ac84eba21492e76dc0c64fc784))
+
+
+
+
+
 ## [0.11.2](https://github.com/washingtonpost/wpds-ui-kit/compare/v0.11.1...v0.11.2) (2022-07-21)